ok
ok

+++ CCGS_WASM/prime_distributed_project/cpp_src/worker_func_5.cpp
@@ -1,9 +1,9 @@
 #include "prime_workers.h"
 
 // Implementation for worker 5.
-uint64_t worker_func_5(uint64_t start, uint64_t end) {
-    uint64_t count = 0;
-    for (uint64_t i = start; i <= end; ++i) {
+uint32_t worker_func_5(uint32_t start, uint32_t end) {
+    uint32_t count = 0;
+    for (uint32_t i = start; i <= end; ++i) {
         if (is_prime(i)) {
             count++;
         }
